TLS enabled universially by importing credentials

+++ ublox-short-range/src/command/data_mode/urc.rs
@@ -10,7 +10,7 @@ use no_std_net::IpAddr;
 #[derive(Debug, PartialEq, Clone, AtatResp)]
 pub struct PeerConnected {
     #[at_arg(position = 0)]
-    pub handle: u8,
+    pub handle: usize,
     #[at_arg(position = 1)]
     pub connection_type: ConnectionType,
     #[at_arg(position = 2)]
